added ffi and bumped to 2.0.2-upd.1

+++ build.sh
@@ -4,6 +4,11 @@ set -e
 DIST_DIR=dist
 export GYP_MSVS_VERSION=2015
 
-\rm -rf ${DIST_DIR}
-./vendors/7z/7z.exe x vendors/electron-*.zip -o${DIST_DIR}
-./tasks/winax/build.sh --update-first
+if [[ ! "$1" == "--fast" ]]; then
+  \rm -rf ${DIST_DIR}
+  ./vendors/7z/7z.exe x vendors/electron-*.zip -o${DIST_DIR}
+fi
+
+for proj in winax ffi; do
+  ./tasks/${proj}/build.sh --update-first
+done
